Worked through the vimtutor. Add .vimrc that works.
So ... all the cool kids are using vim these days. I started out as an
emacs guy so that is what I use. but I want to make sure that I am not
missing out on anything. It seems that even the emacs guys are saying
that you should be using 'evil' mode to do your text editing. So there
must be something to this vim thing right?
At the end of the vim tutor they suggest making a copy of the default
vimrc and making it your own. Consider this my ititial commit of a
legitimate vimrc.