Set default language and EDITOR.
Set the default language to use US english and utf8 incase this helps
anything to be more awesome in the future.
Set EDITOR environment variable to vim, if something on the command line
want to make a quick edit to something, I don't want that to launch my
huge emacs setup.